Improve EntityNotFoundException by adding helper methods
Added two new static methods, `ThrowIfNull` and `Throw`, to the `EntityNotFoundException` class to assist in handling null entity checks and increasing error reporting efficiency. This update widens the class's range of use, offering more detailed error reports which include information about the entity type and key. The main goal of these enhancements is to provide user-friendly error reports that can assist in debugging by providing specific details including the type and key of the entity involved in the exception.